create actual graph and show in main application
